Fix type (ssize_t -> off_t) for gthread-freebsd-sendfile
git-svn-id: svn://svn.lighttpd.net/lighttpd/trunk@2563 152afb58-edef-0310-8abb-c4023f1b3aa9

+++ src/network_gthread_freebsd_sendfile.c
@@ -96,7 +96,7 @@ gpointer network_gthread_freebsd_sendfile_read_thread(gpointer _srv) {
 
 		if ((wj = g_async_queue_timed_pop(inq, &ts))) {
 			/* let's see what we have to stat */
-			ssize_t r;
+			off_t r;
 			off_t offset;
 			size_t toSend;
 			chunk *c;
